Fixed get and save tag

+++ src/tagmanager.cpp
@@ -15,7 +15,12 @@ namespace ddb {
 std::string TagManager::getTag(const std::string& registry) {
     const auto path = this->ddbFolder / DDB_FOLDER / TAGSFILE;
 
+    LOGD << "Path = " << path;
+    LOGD << "Getting tag of registry " << registry;
+
     if (!exists(path)) {
+        LOGD << "Path does not exist, creating empty file";
+
         std::ofstream out(path, std::ios_base::out);
         out << "{}";
         out.close();
@@ -26,15 +31,21 @@ std::string TagManager::getTag(const std::string& registry) {
     json j;
     i >> j;
 
+    LOGD << "Contents: " << j.dump();
+
+    if (!j.contains(registry)) return "";
+
     const auto reg = j[registry];
 
     if (j.is_null()) return "";
 
     return reg["tag"];
-
 }
 void TagManager::setTag(const std::string& tag, const std::string& registry) {
     const auto path = this->ddbFolder / DDB_FOLDER / TAGSFILE;
+
+    LOGD << "Path = " << path;
+    LOGD << "Setting tag '" << tag << "' of registry " << registry;
 
     if (!exists(path)) {
         std::ofstream out(path, std::ios_base::out);
@@ -45,14 +56,19 @@ void TagManager::setTag(const std::string& tag, const std::string& registry) {
     std::ifstream i(path);
     json j;
     i >> j;
+    i.close();
 
-    const auto reg = j[registry];
+    LOGD << "Contents: " << j.dump();
 
     // if (reg.is_null()) {
     j[registry] = {{"tag", tag},
                    {"mtime", std::chrono::system_clock::to_time_t(
                                  std::chrono::system_clock::now())}};
-    fs::remove(path);
+
+    if (exists(path)) {
+        fs::remove(path);
+    }
+
     std::ofstream out(path, std::ios_base::out);
     out << std::setw(4) << j;
     out.close();
